+++ wallets/serializers.py
@@ -17,3 +17,10 @@ class DepositSerializer(serializers.Serializer):
         decimal_places=2,
         min_value=Decimal("0.01")
     )
+
+class WithdrawSerializer(serializers.Serializer):
+    amount = serializers.DecimalField(
+        max_digits=12,
+        decimal_places=2,
+        min_value=Decimal("0.01")
+    )